nexusbackend/meter_reading_api/serializers.py
from  rest_framework import serializers
from meter_reading.models import MeterReading
from meter.models import Meter

class MeterReadingSerializer(serializers.ModelSerializer):
    class Meta:
        model = MeterReading
        fields = '__all__'

class MeterSerializer(serializers.ModelSerializer):
    class Meta:
        model = Meter
        fields = '__all__'        